Moved date and time picking composabled to DatePicker.kt
