+++ app/src/main/java/com/pvp/app/ui/screen/task/DatePicker.kt
@@ -1,0 +1,162 @@
+package com.pvp.app.ui.screen.task
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.IntrinsicSize
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.material3.DisplayMode
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.material3.TimeInput
+import androidx.compose.material3.rememberDatePickerState
+import androidx.compose.material3.rememberTimePickerState
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.window.Dialog
+import androidx.compose.ui.window.DialogProperties
+import java.time.Instant
+import java.time.LocalDateTime
+import java.time.ZoneId
+import java.time.format.DateTimeFormatter
+
+val dateFormatter: DateTimeFormatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm")
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun DatePickerDialog(
+    showPicker: Boolean,
+    onDismiss: () -> Unit,
+    onDateSelected: (LocalDateTime) -> Unit,
+) {
+    val stateDate = rememberDatePickerState(initialDisplayMode = DisplayMode.Input)
+
+    if (showPicker) {
+        androidx.compose.material3.DatePickerDialog(
+            onDismissRequest = onDismiss,
+            confirmButton = {
+                TextButton(
+                    onClick = {
+                        onDismiss()
+
+                        val instant = stateDate.selectedDateMillis?.let {
+                            Instant.ofEpochMilli(it)
+                        }
+
+                        instant?.let {
+                            onDateSelected(LocalDateTime.ofInstant(it, ZoneId.systemDefault()))
+                        }
+                    }
+                ) {
+                    Text("OK")
+                }
+            },
+            dismissButton = {
+                TextButton(onClick = onDismiss) {
+                    Text("CANCEL")
+                }
+            }
+        ) {
+            androidx.compose.material3.DatePicker(state = stateDate)
+        }
+    }
+}
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun TimePickerDialog(
+    showPicker: Boolean,
+    onDismiss: () -> Unit,
+    onTimeSelected: (Int, Int) -> Unit,
+    initialHour: Int,
+    initialMinute: Int
+) {
+    val stateTime = rememberTimePickerState(
+        initialHour = initialHour,
+        initialMinute = initialMinute,
+        is24Hour = true
+    )
+
+    if (showPicker) {
+        TimePicker(
+            onCancel = onDismiss,
+            onConfirm = {
+                onDismiss()
+                onTimeSelected(stateTime.hour, stateTime.minute)
+            },
+        ) {
+            TimeInput(state = stateTime)
+        }
+    }
+}
+@Composable
+fun TimePicker(
+    title: String = "Select Time",
+    onCancel: () -> Unit,
+    onConfirm: () -> Unit,
+    toggle: @Composable () -> Unit = {},
+    content: @Composable () -> Unit,
+) {
+    Dialog(
+        onDismissRequest = onCancel,
+        properties = DialogProperties(
+            usePlatformDefaultWidth = false
+        )
+    ) {
+        Surface(
+            shape = MaterialTheme.shapes.extraLarge,
+            tonalElevation = 6.dp,
+            modifier = Modifier
+                .width(IntrinsicSize.Min)
+                .height(IntrinsicSize.Min)
+                .background(
+                    shape = MaterialTheme.shapes.extraLarge,
+                    color = MaterialTheme.colorScheme.surface
+                )
+        ) {
+            Column(
+                modifier = Modifier.padding(24.dp),
+                horizontalAlignment = Alignment.CenterHorizontally
+            ) {
+                Text(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(bottom = 20.dp),
+                    text = title,
+                    style = MaterialTheme.typography.labelMedium
+                )
+
+                content()
+
+                Row(
+                    modifier = Modifier
+                        .height(40.dp)
+                        .fillMaxWidth()
+                ) {
+                    toggle()
+
+                    Spacer(modifier = Modifier.weight(1f))
+
+                    TextButton(
+                        onClick = onCancel
+                    ) {
+                        Text("Cancel")
+                    }
+
+                    TextButton(
+                        onClick = onConfirm
+                    ) {
+                        Text("OK")
+                    }
+                }
+            }
+        }
+    }
+}
